+++ test-suite/testkit/tests/runtime.rs
@@ -123,6 +123,15 @@ impl Runtime for TestRuntime {
         Ok(())
     }
 
+    fn initiate_resuming_service(
+        &self,
+        _context: ExecutionContext<'_>,
+        _spec: &InstanceSpec,
+        _parameters: Vec<u8>,
+    ) -> Result<(), ExecutionError> {
+        Ok(())
+    }
+
     fn update_service_status(
         &mut self,
         _snapshot: &dyn Snapshot,